+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ...
 
+## [0.3.3] - 2023-05-26
+
+### Fixed in 0.3.3
+
+- Corrected g2config.Load() signature
+- Updated dependencies
+  - google.golang.org/protobuf v1.30.0
+
 ## [0.3.2] - 2023-03-10
 
 ### Fixed in 0.3.2